Fix error in phantom_model module
The sections that are already exhausted in bending are removed from
the shear checkings in order to avoid the non-linear solver that computes
the equilibrium under normal stresses to crash.
When an element has its first section (index 0) exhausted but not
its second one (index 1) only one phantom element is created for
this second section. The problem comes when the corresponding loaded
nodes are computed: the createElements method put only one node tag
in tagNodesToLoad corresponding to this element, but it puts as the
unique item in the list. When the createLoads method tries to get
the node to load for the second section (index 1) the program crashes
due to an index out of range problem, because the list has only one
item.
From now on, for each exhausted section of an element a None item
is added to the tagNodesToLoad list, so the correspondence between
section indexes and tagNodesToLoad items is maintained.