Fix error in computation of the expressions 6.2.a and 6.2.b
In those expressions the concrete area that must be used to compute
the design shear resistance of the member without shear reinforcement
is the product of the effective depth (d) by the the smallest width
of the cross-section in the tensile area (bw). The error has been
fixed and a new verification test has been added.