Implement functions that return nodal tributary quantities in a dictionary
The tributary area of the nodes is cumulative (to avoid discontinuities for
the nodes at the set boundaries), so if, by accident, it is called twice,
you'll get this value doubled unless you call resetTributaries first.
To avoid remembering calling this problem at once I've implemented
three functions that compute the tributary lengths/areas/volumes store
them in a Python dictionary with the node tag as key and call
resetTributaries inmediately. This way the user get the tributary values
he needs for its analysis and there is no risk of cumulating those
values in subsequent calls to the methods that compute them.