Fix error in Action::suma method
Perform the sum only if the given action is compatible. The previous
method of performing the sum and then marking the combination as
incompatible gives poor results. The reason is that the expression marked
as incompatible continues to grow receiving more actions but at the
end is discarded because one of its terms is incompatible with another
one. This has no sense (and discards combinations will be correct except
for one incompatible member) . The new procedure creates only compatible
combinations.