Update code
The addNewLoadCaseToDomain method has changed its behavior. Until now, it
always removed the previous loads and reverted the model to start.
This is counterintuitive given the “add” in its name; one can expect to be
adding loads to the model and not removing the previous ones. From now on
the removal of previous loads and reverting to the initial state is
optional. The reset parameter controls it.