Implement needsUpdate method
This method Return true if the material needs to update its internal
state even if the trial strains have not changed. This is the case
when the material deforms without load or under constant load, for
example by shrinkage or creep. It is used in the update methods
of ForceBeamColumn elements to check if the fiber sections require
to be updates due to the nature of the materials (TDConcrete for
example).