Hide the error message due to a known error
The inertia loads on trusses are simulated by putting two equivalent nodal
loads at their ends. As a consequence, the method that creates the load
returns a null pointer (because there is no ElementolLoad to return).
This is a know problem and, for the time being, the solution is to
redirect the cerr stream to /dev/null.