Create the pier struts and ties in the correct position
Until now those struts and ties were located around the centroid
of the triangle formed by the pile heads regardless of the position
of the pier with respect to those piles. From now on the strut and
ties that represent the pile inside the pile cap are created around
the pier axis.